refactor: 상품 등록, 수정 api 관련 관심사 이동 (#94)
