+++ fe/src/api/product/index.ts
@@ -7,6 +7,7 @@ import {
   Status,
 } from "@api/type";
 import { PRODUCT_API_PATH } from "./constants";
+import { ProductRegisterInfo } from "@components/ProductRegister/type";
 
 export const getAddresses = async (
   page: number = 0,
@@ -38,16 +39,28 @@ export const getProductDetail = async (productId: number) => {
   return data;
 };
 
-export const postProduct = async (productInfo: FormData) => {
+export const postProduct = async (productInfo: ProductRegisterInfo) => {
   const config = {
     headers: {
       "content-type": "multipart/form-data",
     },
   };
 
+  const formData = new FormData();
+  const price = productInfo.price.replace(/,/g, "");
+
+  productInfo.newImages?.forEach((image) => {
+    formData.append("images", image.image);
+  });
+  formData.append("title", productInfo.title);
+  formData.append("content", productInfo.content);
+  formData.append("categoryId", JSON.stringify(productInfo.categoryId));
+  formData.append("addressId", JSON.stringify(productInfo.address.id));
+  formData.append("price", price);
+
   return await fetcher.post<{ productId: number }>(
     PRODUCT_API_PATH.products,
-    productInfo,
+    formData,
     config
   );
 };
@@ -57,7 +70,7 @@ export const patchProduct = async ({
   productInfo,
 }: {
   productId: number;
-  productInfo: FormData;
+  productInfo: ProductRegisterInfo;
 }) => {
   const config = {
     headers: {
@@ -65,9 +78,24 @@ export const patchProduct = async ({
     },
   };
 
+  const formData = new FormData();
+  const price = productInfo.price.replace(/,/g, "");
+
+  productInfo.newImages?.forEach((image) => {
+    formData.append("newImages", image.image);
+  });
+  if (productInfo.deletedImageIds?.length) {
+    formData.append("deletedImageIds", productInfo.deletedImageIds?.join());
+  }
+  formData.append("title", productInfo.title);
+  formData.append("content", productInfo.content);
+  formData.append("categoryId", JSON.stringify(productInfo.categoryId));
+  formData.append("addressId", JSON.stringify(productInfo.address.id));
+  formData.append("price", price);
+
   return await fetcher.patch(
     `${PRODUCT_API_PATH.products}/${productId}`,
-    productInfo,
+    formData,
     config
   );
 };
